More changes to prepare for C++17.

+++ src/function/ROLZCodec.hpp
@@ -51,7 +51,7 @@ namespace kanzi {
 
 		int get();
 
-		void setContext(int ctx) { _ctx = uint8(ctx) << _logSize; }
+		void setContext(byte ctx) { _ctx = uint8(ctx) << _logSize; }
 	};
 
 	class ROLZEncoder {
